rename `training/folding` to `training`
