+++ src/kfold/training/dataset/cropper/__init__.py
