+++ src/kfold/training/dataset/cropper/base.py
