+++ src/kfold/training/dataset/cropper/multi_anchor.py
