+++ src/kfold/training/dataset/filter/base.py
