+++ src/kfold/training/dataset/filter/resolution.py
