+++ src/kfold/training/dataset/filter/size.py
