+++ src/kfold/training/dataset/utils/pre_crop.py
