+++ src/kfold/training/metrics/structure_metrics.py
@@ -3,7 +3,7 @@ from collections.abc import Sequence
 import torch
 
 from kfold.data.model_input import FoldingInput
-from kfold.training.folding.dataset.utils.permutation import get_aligned_true_coords
+from kfold.training.dataset.utils.permutation import get_aligned_true_coords
 from kfold.utils.geometry.rigid_align import rigid_align
