+++ tests/structure_module/test_ecsi_interpolation.py
@@ -17,7 +17,7 @@ from kfold.config import load_config
 from kfold.data.model_input import FoldingInput
 from kfold.data.tokenized import TokenizedStructure
 from kfold.model.modules.structure_module.kfold_ecsi import KFoldECSI
-from kfold.training.folding.dataset.datamodule import TrainingDataModule
+from kfold.training.dataset.datamodule import TrainingDataModule
 from kfold.utils import errors
 from kfold.utils.registry import Registry
